FMCOMMS2 Source: Use integer types for some parameters

+++ include/gnuradio/iio/fmcomms2_source.h
@@ -49,10 +49,10 @@ namespace gr {
        * class. iio::device::make is the public interface for
        * creating new instances.
        */
-      static sptr make(const std::string &host,
-		    double frequency, double samplerate, double bandwidth,
+      static sptr make(const std::string &host, unsigned long long frequency,
+		    unsigned long samplerate, unsigned long bandwidth,
 		    bool ch1_en, bool ch2_en, bool ch3_en, bool ch4_en,
-		    unsigned int buffer_size, bool quadrature, bool rfdc,
+		    unsigned long buffer_size, bool quadrature, bool rfdc,
 		    bool bbdc, const char *gain1, double gain1_value,
 		    const char *gain2, double gain2_value,
 		    const char *rf_port_select);
